refactor: group AbcdeScore component files
