+++ src/components/Mollecules/AbcdeScore/AbcdeScore.vue
